backend/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
openai
Pillow
torch
transformers
pandas
scikit-learn
uvicorn
fastapi
sentence-transformers
langchain_huggingface
python-multipart